requirements.txt
alembic==1.10.3
apispec==6.3.0
async-timeout==4.0.2
autopep8==2.0.2
click==8.1.3
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-smorest==0.41.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
passlib==1.7.4
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
redis==4.5.4
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.2.3
flask-redis